@@ -3,26 +3,26 @@
   "version": "1.22.19",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@craco/craco": "^6.2.0",
+    "@ant-design/icons": "^4.8.0",
+    "@craco/craco": "^7.1.0",
     "@reduxjs/toolkit": "^1.8.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "antd": "^4.19.5",
-    "axios": "^0.26.1",
-    "classnames": "^2.3.1",
-    "express": "^4.17.3",
+    "antd": "^4.24.8",
+    "axios": "^1.3.4",
+    "classnames": "^2.3.2",
+    "craco-antd": "^2.0.0",
     "lodash": "^4.17.21",
-    "moment": "^2.29.1",
+    "moment": "^2.29.4",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18",
+    "react-dom": "18",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.9.0",
+    "react-scripts": "^5.0.1",
     "redux": "^4.1.2",
-    "source-map-explorer": "^2.5.2",
+    "source-map-explorer": "^2.5.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -54,20 +54,21 @@
     "@babel/eslint-parser": "^7.17.0",
     "@babel/preset-react": "^7.16.7",
     "babel-plugin-import": "^1.13.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.37.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-security": "^1.7.1",
     "husky": "^7.0.4",
     "lint-staged": "^12.3.7",
-    "node-sass": "^6.0.1",
+    "node-sass": "^8.0.0",
     "prettier": "^2.6.1",
-    "sass-loader": "^12.6.0"
+    "sass-loader": "^13.2.2"
   },
-  "packageManager": "yarn@1.22.1"
+  "packageManager": "yarn@1.22.19"
 }
