@@ -1,34 +1,24 @@
 import React from 'react';
-import { Switch } from 'react-router-dom';
-import AuthRoute from '@/routes/AuthRoute';
-import GuestRoute from '@/routes/GuestRoute';
+import { Route, Routes } from 'react-router-dom';
 import HomePage from '@/pages/HomePage';
+import { AuthLayout, GuestLayout } from '@/components/Layouts';
 import LoginPage from '@/pages/LoginPage';
 import ProfilePage from '@/pages/ProfilePage';
 import NotFoundPage from '@/pages/NotFoundPage';
 
-function Routes() {
+function Routing() {
   return (
-    <Switch>
-      <GuestRoute
-        exact
-        path='/'
-        component={HomePage}
-      />
-      <GuestRoute
-        path='/login'
-        component={LoginPage}
-      />
-      <AuthRoute
-        path='/me'
-        component={ProfilePage}
-      />
-      <GuestRoute
-        path='*'
-        component={NotFoundPage}
-      />
-    </Switch>
+    <Routes>
+      <Route element={<GuestLayout />}>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/login" element={<LoginPage />} />
+      </Route>
+      <Route element={<AuthLayout />}>
+        <Route path="/me" element={<ProfilePage />} />
+      </Route>
+      <Route path="*" element={<NotFoundPage />} />
+    </Routes>
   );
 }
 
-export default Routes;
+export default Routing;
