@@ -1,13 +1,13 @@
 import React from 'react';
-import Routes from '@/routes';
-import { BrowserRouter as Router } from 'react-router-dom';
+import Routing from '@/routes';
+import { BrowserRouter } from 'react-router-dom';
 import './App.scss';
 
 function App() {
   return (
-    <Router>
-      <Routes />
-    </Router>
+    <BrowserRouter>
+      <Routing />
+    </BrowserRouter>
   );
 }
 
