@@ -1,17 +1,13 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import { Outlet } from 'react-router-dom';
 import './index.scss';
 
-function GuestLayout({ children }) {
+function GuestLayout() {
   return (
     <div className="guest-layout">
-      {children}
+      <Outlet />
     </div>
   );
 }
-
-GuestLayout.propTypes = {
-  children: PropTypes.element,
-};
 
 export default GuestLayout;
